works i guess?
seems to min max the pressures, which does maximize the efficiency of the cycle so...hurrah, i guess. If we are worried about the lifespan of our turbine then we need to lower the pressure of the turbine entrance so that we get better quality steam. Bad quality steam=damage to turbine, and higher pressure=bad quality. also has the start of being able to plot the cycle on a matplotlib plot.